Creating Splitmenu and refactoring DesktopMenu to allow SplitMenu to be inside
